Feature/be/8/나의 기록 날짜별 조회

See merge request s08-bigdata-recom-sub2/S08P22A301!59

Former-commit-id: ae5a150cf5f6bc643cd18d49fb9e8f4998402c21

+++ backend/sanneomeo/src/main/java/com/hikers/sanneomeo/controller/UserController.java
@@ -1,29 +1,25 @@
 package com.hikers.sanneomeo.controller;
 
 
+import static com.hikers.sanneomeo.exception.BaseResponseStatus.FAIL;
 import static com.hikers.sanneomeo.exception.BaseResponseStatus.UNAUTHORIZED_USER;
 
 import com.hikers.sanneomeo.dto.response.BaseResponseDto;
-import com.hikers.sanneomeo.dto.response.ChallengeResponseDto;
+import com.hikers.sanneomeo.dto.response.GetUserPhotosByDateResponseDto;
 import com.hikers.sanneomeo.exception.BaseException;
 import com.hikers.sanneomeo.exception.BaseResponseStatus;
 import com.hikers.sanneomeo.service.UserService;
-import com.hikers.sanneomeo.service.UserServiceImpl;
 import com.hikers.sanneomeo.utils.JwtTokenUtils;
 
-import java.util.HashMap;
+import java.util.Date;
 import java.util.List;
 import java.util.Map;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -71,11 +67,31 @@ public class UserController {
 
 
   //여기서부터 crud -> 위는 그냥 임시용
-  @PutMapping("/photo/{photoSeq}")
-  public BaseResponseDto<Boolean> changePhotoStatus(@PathVariable Long photoSeq){
+
+  @GetMapping("{userSeq}/photo")
+  public BaseResponseDto<Map<String, Map<String, List<String>>>> getUserPhotos(@PathVariable("userSeq") Long userSeq,
+      @RequestParam("month") Integer month){
+    try{
+      return new BaseResponseDto<>(userService.getUserPhotos(userSeq, month));
+    } catch(Exception e){
+      if(e instanceof BaseException){
+        throw e;
+      } else{
+        throw new BaseException(FAIL);
+      }
+    }
+  }
+
+
+  @PutMapping("{userSeq}/photo/{photoSeq}")
+  public BaseResponseDto<Boolean> changePhotoStatus(@PathVariable("userSeq") Long userSeq,
+      @PathVariable("photoSeq") Long photoSeq){
     try {
       //요청 내부의 userSeq와 인증된 userSeq가 다를 경우
       Long authUserSeq = Long.parseLong(SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString());
+      if(userSeq!=authUserSeq){
+        throw new BaseException(UNAUTHORIZED_USER);
+      }
 
       boolean result = userService.updatePhotoStatus(authUserSeq, photoSeq);
 
@@ -90,12 +106,15 @@ public class UserController {
     }
   }
 
-    @DeleteMapping("/photo/{photoSeq}")
-    public BaseResponseDto<Boolean> deletePhoto(@PathVariable Long photoSeq){
+    @DeleteMapping("{userSeq}/photo/{photoSeq}")
+    public BaseResponseDto<Boolean> deletePhoto(@PathVariable("userSeq") Long userSeq, @PathVariable("photoSeq") Long photoSeq){
       try {
         //요청 내부의 userSeq와 인증된 userSeq가 다를 경우 처리 위해 사용
+        //요청 내부의 userSeq와 인증된 userSeq가 다를 경우
         Long authUserSeq = Long.parseLong(SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString());
-
+        if(userSeq!=authUserSeq){
+          throw new BaseException(UNAUTHORIZED_USER);
+        }
         boolean result = userService.deletePhoto(authUserSeq, photoSeq);
 
         return new BaseResponseDto<>(result);
